Ripping out GeneralLedger and renaming to Accounting to prepare for a world of `with Accounting:` so you know when the Accountant is watching your every move
