+++ enron/__init__.py
@@ -1,8 +1,8 @@
-from .enron import Asset, AssetBalance, AssetAmount, Account, \
-    AccountEntry, DoubleEntry, GeneralLedger, ExchangeRate, ExchangeEntry, \
-    AccountGroup, AssetPair, AutoAccountGroup
-
+from enron.enron import Asset, AssetBalance, AssetAmount, Account, \
+    AccountEntry, DoubleEntry, ExchangeRate, ExchangeEntry, \
+        AccountGroup, AssetPair, AutoAccountGroup
+from enron.accounting import Accountant
 
 __all__ = ("Asset", "AssetPair", "AssetAmount", "Account", "AccountGroup",
-           "AccountEntry", "DoubleEntry", "AssetBalance", "GeneralLedger", "ExchangeRate",
+           "AccountEntry", "DoubleEntry", "AssetBalance", "Accountant", "ExchangeRate",
            "ExchangeEntry", "AccountGroup", "AutoAccountGroup")
